requirements.txt
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
emoji==1.2.0
environs==9.3.2
Faker==8.1.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.17.0
six==1.15.0
SQLAlchemy==1.4.9
text-unidecode==1.3
validate-email==1.3
Werkzeug==1.0.1
